requirements.txt
aniso8601==4.1.0
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-JWT-Extended==3.16.0
Flask-RESTful==0.3.7
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
marshmallow==2.18.0
more-itertools==5.0.0
pluggy==0.8.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.7.0
PyJWT==1.7.1
pytest==4.1.0
python-dotenv==0.10.1
pytz==2018.9
six==1.12.0
Werkzeug==0.14.1
